PR + better match for Overlay_Load

+++ include/ultra64/libc.h
@@ -6,7 +6,7 @@
 void osSyncPrintf(const char* fmt, ...);
 
 void bzero(void* __s, int __n);
-int bcmp(const void* __sl, const void* __s2, size_t __n);
-void bcopy(const void* __src, void* __dest, size_t __n);
+int bcmp(const void* __sl, const void* __s2, int __n);
+void bcopy(const void* __src, void* __dest, int __n);
 
 #endif
